+++ frontend/components/screens/schedule/create/tourism/tour-content-type/TourContentType.module.scss
@@ -1,0 +1,27 @@
+.container {
+	@apply mt-10 flex flex-col gap-4;
+
+	// 관광타입 선택
+	.tour_content_type_container {
+		@apply flex gap-4;
+
+		.btn_container {
+			@apply w-1/4;
+		}
+		.description_container {
+			@apply w-3/4 flex items-center;
+		}
+	}
+
+	//서비스 분류 선택
+	.service_category_container {
+		@apply flex gap-4;
+
+		.btn_container {
+			@apply w-1/4;
+		}
+		.description_container {
+			@apply w-3/4 flex items-center;
+		}
+	}
+}
