+++ frontend/components/screens/schedule/create/tourism/tour-content-type/tour-content-type.interface.ts
@@ -1,0 +1,6 @@
+import { PeriodsType } from '@/atoms/periodAtom';
+
+export interface TourContentTypeProps {
+	onChangePeriods: (dates: PeriodsType[]) => void;
+	isSelectedPeriod: PeriodsType;
+}
