+++ frontend/components/screens/schedule/create/tourism/tour-search/TourSearch.module.scss
@@ -1,0 +1,3 @@
+.container {
+	@apply mt-10;
+}
